Move database and Discord init to different files
