+++ api/src/discord/index.ts
@@ -1,9 +1,30 @@
 import { Client, GatewayIntentBits } from "discord.js";
 import config from "../config";
 import ready from "./listeners/ready";
+import { connectToDatabase } from "../utils/database";
 
 let client: Client | null = null;
 export let usingDiscordBot = true;
+
+export async function initDiscordBot() {
+  getDiscordClient();
+
+  const pgClient = await connectToDatabase();
+
+  try {
+    const query =
+      "UPDATE ctfnote.settings SET discord_integration_enabled = $1";
+    const values = [config.discord.use.toLowerCase() !== "false"];
+    await pgClient.query(query, values);
+  } catch (error) {
+    console.error(
+      "Failed to set discord_integration_enabled flag in the database:",
+      error
+    );
+  } finally {
+    pgClient.release();
+  }
+}
 
 export function getDiscordClient(): Client | null {
   if (!usingDiscordBot) return null;
